feat: ⛑️Custom Nether Portal Spread
https://i.imgur.com/2EDSOtU.png

I've completely rewritten the whole **portal spread** mechanics from scratch.
Before it was a separate mod - now it's a Craft Tweaker script.

🧡 The portal has all the same features as before, but better, faster and more advanced:

- ⚡ *Performance is about 100 times better*: less load on TPS, no more lag spikes every 160 ticks
- 📏 *The maximum radius of portals is 4 times bigger*: 64 blocks
- 🆎 *Many more recipes and more varieties of blocks*: now its possible to turn blocks into different kinds of Nether Ores
- 🔁 *Preserving block rotations when transforming*: transforming villages now way cooler
- 🧾 *More detailed chat messages*
- 🔠 *Localization to Ru or Chinese languages*

⚠ Important: [Coal Block]s should now be placed on the corners of the portal.

Details of how the portal works will be described in the Quest Book in the future.

If you are server admin, use `/portal_spread status` command to see all working portals.