fix(config): ☢️Disable NC radiation
NC radiation can be seen on most NuclearCraft items. Its irradiated chunk and player have an indicator in the bottom right corner.

The problem is that this type of radiation caused a notable CPU load (about 1.5% of TPS). Also, it's hard to document this radiation and its behavior, so many beginner players always panic about getting pico-rads.

Lastly, many players on the E2E-E discord voted to remove radiation.

https://i.imgur.com/Mi2KQan.png