feat: 📀Convert seconds in AdvancedRocketry's JEI integrations into ticks
The majority of AdvancedRocketry's JEI recipes display "0 s" as their time, which is not technically true as recipe's duration and is not very helpful in measuring the recipe's duration and energy required for it.
This change adds Bansoukou patch to AdvancedRocketry to change seconds displayed in AdvancedRocketry's JEI integrations into ticks, displaying the actual recipe durations.
https://i.imgur.com/tEegt2a.png